fix(silk): align NSQ arithmetic and resampler with libopus
Critical fixes for SILK encoder NSQ (Noise Shaping Quantization):

- silk_SMULBB: Use signed 16-bit extraction matching libopus
  The previous implementation used unsigned masking (& 0xFFFF) instead
  of signed cast (int16), causing corrupted R-D calculations when
  quantization values were negative.

- silk_INVERSE32_varQ: Port libopus Newton-Raphson implementation
  The simple division approach overflowed when qres=47 and gain=1.0,
  producing negative inverse gains and corrupting the signal chain.

- silk_RSHIFT_ROUND: Fix int32 overflow when adding rounding term
  Use int64 intermediate to handle values near INT32_MAX.

- Add silk_CLZ32 and helper functions for inverse computation

New downsampling resampler (resample_down_fir.go):

- Implement proper AR2+FIR downsampling algorithm for encoder mode
- Use encoder delay matrix (not decoder matrix) for correct timing
- Port FIR coefficients for various downsample ratios (1:2, 1:3, etc.)

Additional improvements:

- noise_shape.go: Port libopus noise shaping parameter computation
- encode_frame.go: Use symmetric float scaling (32768 instead of 32767)
- encoder.go: Track noise shaping state and LTP correlation

These fixes address root causes identified through debug analysis but
SILK quality is still being investigated for remaining issues.

+++ silk/resample_libopus.go
@@ -520,7 +520,7 @@ func (r *LibopusResampler) up2HQ(out []int16, in []int16) {
 		r.sIIR[2] = out32_2 + X
 
 		// Convert back to int16 and store even sample
-		out[2*k] = sat16(rshiftRound(out32_1, 10))
+		out[2*k] = int16(sat16(rshiftRound(out32_1, 10)))
 
 		// First all-pass section for odd output sample
 		Y = in32 - r.sIIR[3]
@@ -541,7 +541,7 @@ func (r *LibopusResampler) up2HQ(out []int16, in []int16) {
 		r.sIIR[5] = out32_2 + X
 
 		// Convert back to int16 and store odd sample
-		out[2*k+1] = sat16(rshiftRound(out32_1, 10))
+		out[2*k+1] = int16(sat16(rshiftRound(out32_1, 10)))
 	}
 }
 
@@ -568,7 +568,7 @@ func (r *LibopusResampler) firInterpol(out []int16, outIdx int, buf []int16, max
 		resQ15 = smlabb(resQ15, int32(buf[bufIdx+7]), int32(silkResamplerFracFIR12[11-tableIndex][0]))
 
 		if outIdx < len(out) {
-			out[outIdx] = sat16(rshiftRound(resQ15, 15))
+			out[outIdx] = int16(sat16(rshiftRound(resQ15, 15)))
 			outIdx++
 		}
 	}
